WebGL2 polyfill cleanup
