+++ modules/webgl2-polyfill/src/assert.js
@@ -1,0 +1,5 @@
+export default function assert(condition, message) {
+  if (!condition) {
+    throw new Error(message || 'webgl2-polyfill: assertion failed.');
+  }
+}
