+++ modules/webgl2-polyfill/src/polyfill-context.js
@@ -8,173 +8,27 @@
 // easy to reuse or repurpose in other projects.
 
 /* eslint-disable camelcase, brace-style */
-import GL from '@luma.gl/constants';
-import {getParameterPolyfill} from './polyfill-get-parameter';
+import assert from './assert';
 
-function assert(condition, message) {
-  if (!condition) {
-    throw new Error(message || 'luma.gl: assertion failed.');
+import {WEBGL2_CONTEXT_POLYFILLS, WEBGL2_CONTEXT_OVERRIDES} from './polyfill-table';
+
+// Registers extensions, polyfills or mock functions for extensions in the polyfills list
+// TODO - remove use of name `luma`.
+export default function polyfillContext(gl) {
+  gl.luma = gl.luma || {};
+  initializeExtensions(gl);
+  if (!gl.luma.polyfilled) {
+    installPolyfills(gl, WEBGL2_CONTEXT_POLYFILLS);
+    installOverrides(gl, {target: gl.luma, target2: gl});
+    gl.luma.polyfilled = true;
   }
+  return gl;
 }
 
-const OES_vertex_array_object = 'OES_vertex_array_object';
-const ANGLE_instanced_arrays = 'ANGLE_instanced_arrays';
-const WEBGL_draw_buffers = 'WEBGL_draw_buffers';
-const EXT_disjoint_timer_query = 'EXT_disjoint_timer_query';
-const EXT_texture_filter_anisotropic = 'EXT_texture_filter_anisotropic';
-
-const ERR_VAO_NOT_SUPPORTED = 'VertexArray requires WebGL2 or OES_vertex_array_object extension';
-
-// Return true if WebGL2 context
-function isWebGL2(gl) {
-  return gl && gl.TEXTURE_BINDING_3D === GL.TEXTURE_BINDING_3D;
-}
-
-// Return object with webgl2 flag and an extension
-function getExtensionData(gl, extension) {
-  return {
-    webgl2: isWebGL2(gl),
-    ext: gl.getExtension(extension)
-  };
-}
-
-// function mapExtensionConstant(gl, constant) {
-//   switch (constant) {
-//   case ext.FRAGMENT_SHADER_DERIVATIVE_HINT_OES: return GL.FRAGMENT_SHADER_DERIVATIVE_HINT;
-//   }
-// }
-
-const WEBGL_CONTEXT_POLYFILLS = {
-  // POLYFILL TABLE
-  [OES_vertex_array_object]: {
-    meta: {suffix: 'OES'},
-    // NEW METHODS
-    createVertexArray: () => {
-      assert(false, ERR_VAO_NOT_SUPPORTED);
-    },
-    deleteVertexArray: () => {},
-    bindVertexArray: () => {},
-    isVertexArray: () => false
-  },
-  [ANGLE_instanced_arrays]: {
-    meta: {
-      suffix: 'ANGLE'
-      // constants: {
-      //   VERTEX_ATTRIB_ARRAY_DIVISOR: 'VERTEX_ATTRIB_ARRAY_DIVISOR_ANGLE'
-      // }
-    },
-    vertexAttribDivisor(location, divisor) {
-      // Accept divisor 0 even if instancing is not supported (0 = no instancing)
-      assert(divisor === 0, 'WebGL instanced rendering not supported');
-    },
-    drawElementsInstanced: () => {},
-    drawArraysInstanced: () => {}
-  },
-  [WEBGL_draw_buffers]: {
-    meta: {
-      suffix: 'WEBGL'
-    },
-    drawBuffers: () => {
-      assert(false);
-    }
-  },
-  [EXT_disjoint_timer_query]: {
-    meta: {suffix: 'EXT'},
-    // WebGL1: Polyfills the WebGL2 Query API
-    createQuery: () => {
-      assert(false);
-    },
-    deleteQuery: () => {
-      assert(false);
-    },
-    beginQuery: () => {
-      assert(false);
-    },
-    endQuery: () => {},
-    getQuery(handle, pname) {
-      return this.getQueryObject(handle, pname);
-    },
-    // The WebGL1 extension uses getQueryObject rather then getQueryParameter
-    getQueryParameter(handle, pname) {
-      return this.getQueryObject(handle, pname);
-    },
-    getQueryObject: () => {}
-  },
-  OVERRIDES: {
-    // Ensure readBuffer is a no-op
-    readBuffer: (gl, originalFunc, attachment) => {
-      if (isWebGL2(gl)) {
-        originalFunc(attachment);
-      } else {
-        // assert(attachment !== GL_COLOR_ATTACHMENT0 && attachment !== GL_FRONT);
-      }
-    },
-    // Override for getVertexAttrib that returns sane values for non-WebGL1 constants
-    getVertexAttrib: (gl, originalFunc, location, pname) => {
-      // const gl = this; // eslint-disable-line
-      const {webgl2, ext} = getExtensionData(gl, ANGLE_instanced_arrays);
-
-      let result;
-      switch (pname) {
-        // WebGL1 attributes will never be integer
-        case GL.VERTEX_ATTRIB_ARRAY_INTEGER:
-          result = !webgl2 ? false : undefined;
-          break;
-        // if instancing is not available, return 0 meaning divisor has not been set
-        case GL.VERTEX_ATTRIB_ARRAY_DIVISOR:
-          result = !webgl2 && !ext ? 0 : undefined;
-          break;
-        default:
-      }
-
-      return result !== undefined ? result : originalFunc(location, pname);
-    },
-    // Handle transform feedback and uniform block queries in WebGL1
-    getProgramParameter: (gl, originalFunc, program, pname) => {
-      if (!isWebGL2(gl)) {
-        switch (pname) {
-          case GL.TRANSFORM_FEEDBACK_BUFFER_MODE:
-            return GL.SEPARATE_ATTRIBS;
-          case GL.TRANSFORM_FEEDBACK_VARYINGS:
-            return 0;
-          case GL.ACTIVE_UNIFORM_BLOCKS:
-            return 0;
-          default:
-        }
-      }
-      return originalFunc(program, pname);
-    },
-    getInternalformatParameter: (gl, originalFunc, target, format, pname) => {
-      if (!isWebGL2(gl)) {
-        switch (pname) {
-          case GL.SAMPLES:
-            return new Int32Array([0]);
-          default:
-        }
-      }
-      return gl.getInternalformatParameter(target, format, pname);
-    },
-    getTexParameter(gl, originalFunc, target, pname) {
-      switch (pname) {
-        case GL.TEXTURE_MAX_ANISOTROPY_EXT:
-          const {extensions} = gl.luma;
-          const ext = extensions[EXT_texture_filter_anisotropic];
-          pname = (ext && ext.TEXTURE_MAX_ANISOTROPY_EXT) || GL.TEXTURE_MAX_ANISOTROPY_EXT;
-          break;
-        default:
-      }
-      return originalFunc(target, pname);
-    },
-    getParameter: getParameterPolyfill,
-    hint(gl, originalFunc, pname, value) {
-      // TODO - handle GL.FRAGMENT_SHADER_DERIVATIVE_HINT:
-      // switch (pname) {
-      // case GL.FRAGMENT_SHADER_DERIVATIVE_HINT:
-      // }
-      return originalFunc(pname, value);
-    }
-  }
-};
+// TODO - is this still required?
+/* global window, global */
+const global_ = typeof global !== 'undefined' ? global : window;
+global_.polyfillContext = polyfillContext;
 
 function initializeExtensions(gl) {
   gl.luma.extensions = {};
@@ -185,9 +39,30 @@ function initializeExtensions(gl) {
   }
 }
 
+// Install simple overrides (mostly get* functions)
+function installOverrides(gl, {target, target2}) {
+  Object.keys(WEBGL2_CONTEXT_OVERRIDES).forEach(key => {
+    if (typeof WEBGL2_CONTEXT_OVERRIDES[key] === 'function') {
+      // install an override, if no implementation was detected
+      const originalFunc = gl[key] ? gl[key].bind(gl) : () => {};
+      const polyfill = WEBGL2_CONTEXT_OVERRIDES[key].bind(null, gl, originalFunc);
+      target[key] = polyfill;
+      target2[key] = polyfill;
+    }
+  });
+}
+
+function installPolyfills(gl, polyfills) {
+  for (const extension of Object.getOwnPropertyNames(polyfills)) {
+    if (extension !== 'overrides') {
+      polyfillExtension(gl, {extension, target: gl.luma, target2: gl});
+    }
+  }
+}
+
 // Polyfills a single WebGL extension into the `target` object
 function polyfillExtension(gl, {extension, target, target2}) {
-  const defaults = WEBGL_CONTEXT_POLYFILLS[extension];
+  const defaults = WEBGL2_CONTEXT_POLYFILLS[extension];
   assert(defaults);
 
   const {meta = {}} = defaults;
@@ -195,7 +70,7 @@ function polyfillExtension(gl, {extension, target, target2}) {
 
   const ext = gl.getExtension(extension);
 
-  Object.keys(defaults).forEach(key => {
+  for (const key of Object.keys(defaults)) {
     const extKey = `${key}${suffix}`;
 
     let polyfill = null;
@@ -215,40 +90,5 @@ function polyfillExtension(gl, {extension, target, target2}) {
       target[key] = polyfill;
       target2[key] = polyfill;
     }
-  });
-}
-
-// Install simple overrides (mostly get* functions)
-function installOverrides(gl, {target, target2}) {
-  const {OVERRIDES} = WEBGL_CONTEXT_POLYFILLS;
-  Object.keys(OVERRIDES).forEach(key => {
-    if (typeof OVERRIDES[key] === 'function') {
-      // install an override, if no implementation was detected
-      const originalFunc = gl[key] ? gl[key].bind(gl) : () => {};
-      const polyfill = OVERRIDES[key].bind(null, gl, originalFunc);
-      target[key] = polyfill;
-      target2[key] = polyfill;
-    }
-  });
-}
-
-// Registers polyfill or mock functions for all known extensions
-export default function polyfillContext(gl) {
-  // polyfillVertexArrayObject(WebGLRenderingContext, gl);
-  gl.luma = gl.luma || {};
-  initializeExtensions(gl);
-  if (!gl.luma.polyfilled) {
-    Object.getOwnPropertyNames(WEBGL_CONTEXT_POLYFILLS).forEach(extension => {
-      if (extension !== 'overrides') {
-        polyfillExtension(gl, {extension, target: gl.luma, target2: gl});
-      }
-    });
-    installOverrides(gl, {target: gl.luma, target2: gl});
-    gl.luma.polyfilled = true;
   }
-  return gl;
 }
-
-/* global window, global */
-const global_ = typeof global !== 'undefined' ? global : window;
-global_.polyfillContext = polyfillContext;
