+++ modules/webgl2-polyfill/test/index.js
@@ -1,2 +1,2 @@
+import './polyfills/get-parameter-polyfill.spec';
 import './polyfill-context.spec';
-import './polyfill-get-parameter.spec';
